requirements.txt
# Core dependencies
playwright>=1.50.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
httpx==0.25.2
aiofiles==23.2.1

# Date parsing
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0
